backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2


pip install Flask==3.0.0
pip install Flask-Cors==4.0.0
pip install psycopg2-binary==2.9.9
pip install python-dotenv==1.0.0
pip install bcrypt==4.1.2